+++ src/dipdup/__init__.py
@@ -4,6 +4,5 @@ Contains software versions and other metadata.
 """
 import importlib.metadata
 
-# NOTE: Load version from package
 __version__ = importlib.metadata.version('dipdup')
 __spec_version__ = '2.0'