+++ src/dipdup/indexes/tezos_tzkt_operations/matcher.py
@@ -56,8 +56,8 @@ def prepare_operation_handler_args(
                 args.append(operation_data)
                 continue
 
-            if not operation_data.parameter_json:
-                raise FrameworkException('Transaction parameter is empty')
+            if operation_data.parameter_json is None:
+                raise FrameworkException('Processing typed transaction, but parameter_json is None')
             typename = pattern_config.typed_contract.module_name
             type_ = get_parameter_type(package, typename, pattern_config.entrypoint)
             parameter = parse_object(type_, operation_data.parameter_json) if type_ else None