+++ src/dipdup/models/tezos_tzkt.py
@@ -515,6 +515,9 @@ class TzktTokenTransferData(HasLevel):
         to_json = token_transfer_json.get('to') or {}
         standard = token_json.get('standard')
         metadata = token_json.get('metadata')
+        amount = token_transfer_json.get('amount')
+        amount = int(amount) if amount is not None else None
+
         return TzktTokenTransferData(
             id=token_transfer_json['id'],
             level=token_transfer_json['level'],
@@ -529,7 +532,7 @@ class TzktTokenTransferData(HasLevel):
             from_address=from_json.get('address'),
             to_alias=to_json.get('alias'),
             to_address=to_json.get('address'),
-            amount=token_transfer_json.get('amount'),
+            amount=amount,
             tzkt_transaction_id=token_transfer_json.get('transactionId'),
             tzkt_origination_id=token_transfer_json.get('originationId'),
             tzkt_migration_id=token_transfer_json.get('migrationId'),